# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChatterPay Global API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List withdrawal requests
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: requestWithdrawal
      tags: [withdrawals]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalBody"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /withdrawals/{withdrawalId}/approve:
    post:
      operationId: approveWithdrawal
      tags: [withdrawals]
      summary: Approve a withdrawal request (admin only)
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /withdrawals/{withdrawalId}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [withdrawals]
      summary: Reject a withdrawal request (admin only)
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        currency:
          type: string
          enum: [KSH, USDT]
        phone:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
        username:
          type: string
      required: [id, userId, amount, currency, status, createdAt]
    WithdrawalBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
        currency:
          type: string
          enum: [KSH, USDT]
        phone:
          type: string
      required: [amount, currency]
...